Fechamento da versão 1.0

+++ src/test/java/br/ufsc/ine/ppgcc/population/WeightTest.java
@@ -2,6 +2,8 @@ package br.ufsc.ine.ppgcc.population;
 
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class WeightTest {
@@ -9,7 +11,7 @@ class WeightTest {
     @Test
     void constructor() {
         int[] input = {0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0};
-        double expected = 404216;
+        BigDecimal expected = BigDecimal.valueOf(404216);
 
         Weight actual = new Weight(input);
